Changed DjanJinja's csrf protection from a bundle to a Jinja2 extension. This creates {% csrf_token %} as a tag in the environment so it can be used in the same way as Django's own csrf_token template tag.
This lowers the barrier for developers that use both Django's template language and Jinja2 in the same project (admin / front-end, for example) by providing a consistent way to protect their pages.
It also reduces the burden of documentation since Django's own documentation describes exactly how to protect pages in a project using DjanJinja.

It is also tidier - since Django adds a csrf_token context variable, it was nto possible to use a DjanJinja bundle to create a function with the same name.
Template tags are in a different namespace so there is no clash now.